Cleaning branch
ade this branch more bot friendly and less of react app
