--- src/JSON/schema.JSON
@@ -1,165 +1,0 @@
-{
-  "personality-schema": {
-    "$id": "http://localhost/personality",
-    "title": "Personalty",
-    "description": "An object personality qualities that an NPC can have.",
-    "type": "object",
-    "properties": {
-      "genderDescriptors": {
-        "type": "array",
-        "items": {
-          "type": "string"
-        }
-      },
-      "genderRoots": {
-        "type": "array",
-        "items": {
-          "$ref": "http://localhost/personality#/definitions/genderRoot"
-        }
-      },
-      "pronouns": {
-        "type": "array",
-        "items": {
-          "$ref": "http://localhost/personality#/definitions/pronoun"
-        }
-      },
-      "asethtics": {
-        "type": "array",
-        "items": {
-          "type": "string"
-        }
-      },
-      "attractionRoots": {
-        "type": "array",
-        "items": {
-          "type": "string"
-        }
-      },
-      "attractionDescriptors": {
-        "type": "array",
-        "items": {
-          "type": "string"
-        }
-      },
-      "personalityTraits": {
-        "type": "array",
-        "items": {
-          "type": "string"
-        }
-      },
-      "virtues": {
-        "type": "array",
-        "items": {
-          "$ref": "http://localhost/personality#/definitions/virtue"
-        }
-      },
-      "troubles": {
-        "type": "array",
-        "items": {
-          "type": "string"
-        }
-      }
-    },
-    "required": [
-      "genderDescriptors",
-      "genderRoots",
-      "pronouns",
-      "attractionRoots",
-      "attractionDescriptors",
-      "personalityTraits",
-      "virtues",
-      "troubles"
-    ],
-    "definitions": {
-      "genderRoot": {
-        "title": "Gener Root",
-        "description": "A Root Gender.",
-        "type": "object",
-        "properties": {
-          "gen": {
-            "title": "Gender",
-            "description": "The Name of the Root Gender.",
-            "type": "string"
-          },
-          "pron": {
-            "title": "Pronoun (Typical)",
-            "description": "The typical prnoun associated with the Root Gender, not necisarally the given pronoun.",
-            "type": "string"
-          }
-        }
-      },
-      "pronoun": {
-        "title": "Pronoun",
-        "description": "A pronoun, and it's breakdown.",
-        "type": "object",
-        "properties": {
-          "key": {
-            "title": "Prnoun Key",
-            "description": "The Key of the Pronoun, typically the Subject Prnoun.",
-            "type": "string"
-          },
-          "name": {
-            "title": "Prnoun Name",
-            "description": "The Name of the Pronoun, typically the Subject Pronoun / Object Pronoun.",
-            "type": "string"
-          },
-          "subPrn": {
-            "title": "Subject Pronoun",
-            "type": "string"
-          },
-          "objPrn": {
-            "title": "Object Pronoun",
-            "type": "string"
-          },
-          "possAdj": {
-            "title": "Possesive Adjective",
-            "type": "string"
-          },
-          "possPrn": {
-            "title": "Possesive Pronoun",
-            "type": "string"
-          },
-          "rflxPrn": {
-            "title": "Reflexive Pronoun",
-            "type": "string"
-          }
-        },
-        "required": [
-          "key",
-          "name",
-          "subPrn",
-          "objPrn",
-          "possAdj",
-          "possPrn",
-          "rflxPrn"
-        ]
-      },
-      "virtue": {
-        "title": "virtue",
-        "description": "A two part virtue. A Word, and Phrase.",
-        "type": "object",
-        "properties": {
-          "name": {
-            "title": "virtue Name",
-            "description": "The Name of the virtue.",
-            "type": "string"
-          },
-          "description": {
-            "title": "description",
-            "description": "A brief description of the virtue.",
-            "type": "string"
-          }
-        }
-      }
-    }
-  },
-  "names-schema": {
-    "$id": "http://localhost/names",
-    "title": "Character Names",
-    "description": "A collection of thousands of character names",
-    "type": "array",
-    "items": {
-      "type": "string"
-    }
-  }
-}
